requirements.txt
asgiref==3.7.2
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distlib==0.3.8
Django==5.0.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
filelock==3.13.1
idna==3.6
inflection==0.5.1
isodate==0.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
